Changing `/{blog}/tag/*` to `/{blog}/{tag}/*` for more flexability.
Not everyone wants to call a tag a tag. Even if a tag really is just
a tag, it can also be called something else, such as a _category_. I'm
not sure how usefuly this will be to anyone else, or if this is the
proper way to execute this. It works, and we're going to be using it
for work.